Started enabling setting random 'seed's for each individual exercise in an exam, to enable custom exercise selection (e.g., when only putting together a single exam). Currently, only supported in xexams(), exams2pdf(), and exams2html().
In addition to the vector or list specification of the exercise 'file'
in the exams2xyz() interfaces, 'file' can now also be a matrix. This
enables customization of the exact selection of exercises in each
exam. The 'seed' can be a matrix of the same dimension. Currently,
only supported in  xexams(), exams2pdf(), and exams2html().


git-svn-id: svn://r-forge.r-project.org/svnroot/exams/pkg/exams@1706 edb9625f-4e0d-4859-8d74-9fd3b1da38cb